# swagger-user.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ember API 서버",
...
    "version": "v2.2"
  },
  "servers": [
    {
      "url": "https://ember-app.duckdns.org",
      "description": "Ember 배포 서버"
    }
  ],
...
  "paths": {
...
    "/api/dev/exchange-rooms/{roomId}/set-deadline": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "교환일기 마감 시간 변경",
        "description": "교환일기 방의 deadlineAt을 현재 시각 + N분으로 변경. 만료 테스트용.",
        "operationId": "setDeadline",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/dev/exchange-rooms/{roomId}/force-complete": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "교환일기 강제 완주",
        "description": "교환일기 방을 4턴 완료 상태로 강제 변경. 관계 확장 테스트용.",
        "operationId": "forceComplete",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/dev/ai/simulate/{diaryId}": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "AI 분석 결과 시뮬레이션",
        "description": "FastAPI 없이 AI 파이프라인 테스트. RabbitMQ에 Mock 분석 결과 발행 → 2~3초 후 diary_keywords에 감정/성격/라이프스타일/톤 태그 저장.",
        "operationId": "simulateAiResult",
        "parameters": [
          {
            "name": "diaryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dev/token": {
      "get": {
        "tags": [
          "Dev"
        ],
        "summary": "테스트 토큰 발급",
        "description": "카카오 로그인 없이 userId로 JWT accessToken 발급. 프론트 개발/테스트용.",
        "operationId": "issueTestToken",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "토큰을 발급할 사용자 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dev/redis/user/{userId}": {
      "get": {
        "tags": [
          "Dev"
        ],
        "summary": "유저별 Redis 캐시 현황",
        "description": "userId 기준으로 AI 분석, 매칭 추천, RT, 채팅 시퀀스 등 관련 캐시 키 현황 조회.",
        "operationId": "redisUserKeys",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dev/redis/keys": {
      "get": {
        "tags": [
          "Dev"
        ],
        "summary": "Redis 키 패턴 검색",
        "description": "패턴으로 Redis 키 검색. 예: AI:DIARY:*, MATCHING:RECO:*, MSG:SEQ:*",
        "operationId": "redisKeys",
        "parameters": [
          {
            "name": "pattern",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/dev/redis/get": {
      "get": {
        "tags": [
          "Dev"
        ],
        "summary": "Redis 키 조회",
        "description": "특정 키의 값 + TTL(초) 조회.",
        "operationId": "redisGet",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dev/redis/delete": {
      "delete": {
        "tags": [
          "Dev"
        ],
        "summary": "Redis 키 삭제",
        "description": "특정 Redis 키 삭제. 캐시 무효화 테스트용.",
        "operationId": "redisDelete",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}